fix(queue): decouple cap sweep into 10s job (#1319)
The trim sweep was bundled inside queueMonitoring on a 120s cadence. Under
a flood from cli-light direct-Redis enqueues, low-tier workspaces could
overshoot the cap by 5× before the next sweep ran (observed: 900+ waiting
jobs against a 200 cap during 2026-05-26 incident).

Extract the sweep into a dedicated queueCapSweep job scheduled at 10s.
queueMonitoring keeps the heavier alerting + legacy-cleanup paths on its
existing 120s cadence — only the lightweight Lua scan + trim is moved.
